Display images in Flipboard
This commit adapts the Flipboard class to display full images across
its grid of FlipPads. It does this by creating a series of SVG patterns
given an image, one for each tile in the FlipPad grid, which displays a
particular segment. Each FlipPad references its corresponding pattern
using a unique element ID.

A new class, 'SvgImageGrid', handles creating the patterns that
effectively segment the image into a grid of tiles, and making
them available in an SVG container element.