Add TransitionComposer describing transition pattern
Creating this 'grid' of timings quickly becomes complex once
we start trying to provide configuration options to client code.
So we don't bloat Flipboard with ancillary logic, and so we're not
adding more to its constructor parameters, the new TransitionComposer
class will hold this stuff.

Now, all Flipboard has to do to know what delay to apply to its
FlipPad's transitions is ask its TransitionComposer member, giving
a row and col index.

At the moment its config options are fairly basic, in future it'd be
nice to allow more complex parameters, see the class' TODOs.