Use absolute units for <pattern>s
Using relative units, i.e. leaving 'patternUnits' to its default of
'objectBoundingBox', meant the pattern's coorinates were relative to
its bounding box when applied to an SVG. Because of their shape, the
bounding boxes for the bottom flap SVGs are not perfectly aligned
with their containing tile, i.e. the bounding box's (0, 0) did not
match the tile's (0, 0). This is more noticeable with increasing
corner radius, as the bottom flap's left- and top-most points move
further away from x=0 and y=0 respectively.

This meant the bottom and top flap's images were mis-aligned.
Setting 'patternUnits' to 'userSpaceOnUse' means our pattern
coordinates are absolute, and not relative to the bounding box,
so fixing the issue.