Generalise Flipboard's image-setting methods
All we're doing when applying an image to the FlipPad SVGs is setting
their 'fill' attributes, so it makes sense to generalise Flipboard's
'setCurrentImage' and 'setNextImage' methods to accept an object that
provides an arbitrary fill string, rather than having them hard-coded
to just work with SvgImageGrid and its <pattern> IDs.

These methods are now 'setCurrentFill' and 'setNextFill', and accept any
object that implements a 'getFillString(col, row)' function.
SvgImageGrid implemenets this, providing the 'url(#id)' strings that
referencing the corresponding <pattern> elements, but we could also
create e.g. a more simple ColorFill class that simply provides a hex
color.